+++ skbio/sequence/tests/test_substitution.py
@@ -39,8 +39,8 @@ class TestSubstitutionMatrix(TestCase):
         self.assertTupleEqual(obs.shape, (5, 5))
         assert_array_equal(obs.scores, self.scores)
 
-        # data type becomes float
-        self.assertEqual(obs.dtype, np.float64)
+        # data type becomes float32
+        self.assertEqual(obs.dtype, np.float32)
 
         # scores is an alias of data
         assert_array_equal(obs.scores, obs.data)
@@ -99,6 +99,21 @@ class TestSubstitutionMatrix(TestCase):
         obs = SubstitutionMatrix('ACGT', [-1] * 6)
         assert_array_equal(obs.scores, np.identity(4) - 1)
 
+    def test_init_alt_float(self):
+        # alternative floating-point precision (64-bit)
+        obs = SubstitutionMatrix(self.alphabet, self.scores.astype(np.float32))
+        assert_array_equal(obs.scores, self.scores)
+        self.assertEqual(obs.dtype, np.float32)
+        obs = SubstitutionMatrix(self.alphabet, self.scores.astype(np.float64))
+        assert_array_equal(obs.scores, self.scores)
+        self.assertEqual(obs.dtype, np.float64)
+        obs = SubstitutionMatrix(self.alphabet, self.scores.astype(np.float16))
+        assert_array_equal(obs.scores, self.scores)
+        self.assertEqual(obs.dtype, np.float32)
+        obs = SubstitutionMatrix(self.alphabet, self.scores.astype(np.int16))
+        assert_array_equal(obs.scores, self.scores)
+        self.assertEqual(obs.dtype, np.float32)
+
     def test_to_dict(self):
         mat = SubstitutionMatrix(self.alphabet, self.scores)
         obs = mat.to_dict()
@@ -120,6 +135,13 @@ class TestSubstitutionMatrix(TestCase):
                         [0., 1., 0.],
                         [0., 0., 1.]])
         assert_array_equal(obs.data, exp)
+        self.assertEqual(obs._data.dtype, np.float32)
+
+        # data type
+        obs = SubstitutionMatrix.from_dict(d, dtype="float32")
+        self.assertEqual(obs._data.dtype, np.float32)
+        obs = SubstitutionMatrix.from_dict(d, dtype="float64")
+        self.assertEqual(obs._data.dtype, np.float64)
 
         # alphabet is inconsistent
         msg = ('The outer and inner layers of the dictionary must have the '
@@ -137,9 +159,6 @@ class TestSubstitutionMatrix(TestCase):
         d['a']['b'] = 'hello'
         with self.assertRaises(ValueError):
             SubstitutionMatrix.from_dict(d)
-        d['a']['b'] = None
-        with self.assertRaises(TypeError):
-            SubstitutionMatrix.from_dict(d)
 
     def test_identity(self):
         obs = SubstitutionMatrix.identity('ACGT', 1, -2)
@@ -150,11 +169,24 @@ class TestSubstitutionMatrix(TestCase):
                         [-2., -2., 1., -2.],
                         [-2., -2., -2., 1.]])
         assert_array_equal(obs.scores, exp)
+        self.assertEqual(obs._data.dtype, np.float32)
+
+        obs = SubstitutionMatrix.identity('ACGT', 1, -2, dtype="float64")
+        assert_array_equal(obs.scores, exp)
+        self.assertEqual(obs._data.dtype, np.float64)
+
+        obs = SubstitutionMatrix.identity('ACGT', 1, -2, dtype="int16")
+        assert_array_equal(obs.scores, exp)
+        self.assertEqual(obs._data.dtype, np.float32)
+
+        with self.assertRaises(TypeError):
+            _ = SubstitutionMatrix.identity('ACGT', 1, -2, dtype="hello")
 
     def test_by_name(self):
         obs = SubstitutionMatrix.by_name('NUC.4.4')
         self.assertEqual(len(obs.alphabet), 15)
         self.assertEqual(obs['A', 'T'], -4)
+        self.assertEqual(obs._data.dtype, np.float32)
         obs = SubstitutionMatrix.by_name('BLOSUM50')
         self.assertEqual(len(obs.alphabet), 24)
         self.assertEqual(obs['M', 'K'], -2)